Clean up WaitForSessionFile logic and support increasing timeout with warning (#2653)
* Add better sessionfile wait logic with setting

* clean up logic

* move into process.ts to use vscode api

* use timeout seconds

+++ src/process.ts
@@ -17,6 +17,10 @@ export class PowerShellProcess {
         return pspath.replace(new RegExp("'", "g"), "''");
     }
 
+    // This is used to warn the user that the extension is taking longer than expected to startup.
+    // After the 15th try we've hit 30 seconds and should warn.
+    private static warnUserThreshold = 15;
+
     public onExited: vscode.Event<void>;
     private onExitedEmitter = new vscode.EventEmitter<void>();
 
@@ -174,20 +178,36 @@ export class PowerShellProcess {
         return true;
     }
 
-    private waitForSessionFile(): Promise<utils.IEditorServicesSessionDetails> {
-        return new Promise((resolve, reject) => {
-            utils.waitForSessionFile(this.sessionFilePath, (sessionDetails, error) => {
-                utils.deleteSessionFile(this.sessionFilePath);
+    private sleep(ms: number) {
+        return new Promise(resolve => setTimeout(resolve, ms));
+    }
 
-                if (error) {
-                    this.log.write(`Error occurred retrieving session file:\n${error}`);
-                    return reject(error);
-                }
+    private async waitForSessionFile(): Promise<utils.IEditorServicesSessionDetails> {
+        // Determine how many tries by dividing by 2000 thus checking every 2 seconds.
+        const numOfTries = this.sessionSettings.developer.waitForSessionFileTimeoutSeconds / 2;
+        const warnAt = numOfTries - PowerShellProcess.warnUserThreshold;
 
+        // Check every 2 seconds
+        for (let i = numOfTries; i > 0; i--) {
+            if (utils.checkIfFileExists(this.sessionFilePath)) {
                 this.log.write("Session file found");
-                resolve(sessionDetails);
-            });
-        });
+                const sessionDetails = utils.readSessionFile(this.sessionFilePath);
+                utils.deleteSessionFile(this.sessionFilePath);
+                return sessionDetails;
+            }
+
+            if (warnAt === i) {
+                vscode.window.showWarningMessage(`Loading the PowerShell extension is taking longer than expected.
+    If you're using privilege enforcement software, this can affect start up performance.`);
+            }
+
+            // Wait a bit and try again
+            await this.sleep(2000);
+        }
+
+        const err = "Timed out waiting for session file to appear.";
+        this.log.write(err);
+        throw new Error(err);
     }
 
     private onTerminalClose(terminal: vscode.Terminal) {
